requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
python-decouple==3.8
django-cors-headers==4.3.0
django-filter==23.5 